Add 7.2 cluster compatibility
Updated the daemon so that it can detect the version of the SDK and the
version of each cluster that is to be monitored, so that it can support
a mix of 8.0 and 7.2 clusters.

Also, updated the setup_venv.sh so that it is easier to install
different versions of the SDK into a virtual environment

And updated the README.md to reflect these changes.