Added fancy new multi-cluster Grafana dashboards.
In order to more easily support these new dashboards I changed how the
fields are stored in InfluxDB. Instead of '_' for a separator in
measurement, tag, and field names I use '.' because this allows
Grafana's variable substitution to work correctly. Also, instead of
storing a different measurement for each element of a dict stat, I
store them as multiple fields in a single measurement, which is more
efficient for InfluxDB.

Also, updated the influxdb_plugin.py so that it forces int and long
values to float so that we don't run into the unsigned 64 bit
problem and/or the switching data type problem.

Also, I beefed up some of the error handling in the initial
configuration stage so that it spits out useful error messages instead
of just dieing with an exception.

Also, I fixed a bug with configuring the daemon to query for more than
200 different stat keys.

Also, I fixed a bug where if you only have the 8.0 SDK installed and
tried to run the Connector on a 7.2 cluster it would not work. Now it
does.

Also, I tried to make it a little more straight forward to use the
Connector by making the local config file be the default. That way you
can easily start and stop it with simple ./isi_data_insights_d.py start
and stop commands, no need to specify the --config param anymore.

Also, fixed bug with specifying multiple clusters on multiple lines in
the config file.

Also, updated requirements.txt to include the pip installable Isilon
SDKs.

Lastly, I improved README.md install and setup instructions.