Added cluster composite, equation, and pct change stats.
Added logic for deriving stats by compositing node stats, computing a
custom equations, computing percent change in stats, and finally
computing custom equations from any of the previously derived stats.

Includes ability to address specific fields of stat value for derived
stats.  If the value of a particular stat key is a dict or list then
you can now use a specific field or index in a derived stat computation.
Note that if the value is a list with only a single dict element then
you can address a field of the dict directly with the field name or by
using the 0 index.

Added config file parsing for the derived stats.

And added Equations to the requirements.txt.