Force the stats to be named as they are in the config file
Previously, if the stat name in the config file was an alias then it
would get re-named back to the "real" name at config time. Now it keeps
its original name from the config file regardless of whether or not it
is an alias.

This means that if you upgrade to this revision and you are using the
provided example_isi_data_insights_d.cfg as your config file and the
provided Grafana dashboards, then you need to change the name of the
node.clientstats.active.nfs3 stat to node.clientstats.active.nfs because
even though it was named "nfs3" it was actually getting inserted into
InfluxDB as "nfs" because that is the "real" name of that particular
stat. So in order for your data to be continuous, you'll need to fix
your config file (see the updates to example_isi_data_insights_d.cfg
for reference).