Enable parallel queries by leveraging gevent
Minimal changes to allow all of the stat queries for multiple clusters
to execute in parallel. Obviously the GIL still means that there can
only be one active thread but since the data collector spends most of
the time waiting for a response from the various clusters or sleeping,
this simplistic approach actually works well in practice.

Internally at Isilon, it's easily able to keep up with monitoring the
statistics for 24 separate clusters every 30 seconds.